feat: update CruiseDays property to float and show up to 2 digits

+++ backend/ResearchCruiseApp/Application/Models/DTOs/CruiseApplications/CruiseApplicationDto.cs
@@ -42,7 +42,7 @@ public class CruiseApplicationDto
 
     public string? CruiseHours { get; init; }
 
-    public int? CruiseDays { get; set; }
+    public float? CruiseDays { get; set; }
 
     public string? AcceptablePeriodBeg { get; init; }
